move endpoint to new package
