+++ pkg/endpoint/endpoint_test.go
@@ -1,4 +1,4 @@
-package client
+package endpoint
 
 import (
 	"testing"
